split up the tree module a little bit more
