+++ src/lib.rs
@@ -39,7 +39,7 @@ extern crate cpuprofiler;
 extern crate rand;
 
 /// atomic lock-free tree
-pub use tree::{Tree, TreeIter};
+pub use tree::{Iter, Tree};
 /// lock-free pagecache
 #[doc(hidden)]
 pub use ds::{Radix, Stack};
@@ -118,4 +118,19 @@ fn uptime() -> std::time::Duration {
     }
 
     START.elapsed()
+}
+
+// This function is useful for inducing random jitter into our atomic
+// operations, shaking out more possible interleavings quickly. It gets
+// fully elliminated by the compiler in non-test code.
+#[inline(always)]
+fn debug_delay() {
+    #[cfg(test)]
+    {
+        use rand::{Rng, thread_rng};
+
+        if thread_rng().gen_weighted_bool(1000) {
+            std::thread::sleep(std::time::Duration::from_millis(100));
+        }
+    }
 }
